Adapt test infrastructure to new DAG storage API

+++ block-storage/src/test/scala/coop/rchain/blockstorage/InMemBlockDagStorage.scala
@@ -33,8 +33,8 @@ final class InMemBlockDagStorage[F[_]: Monad](
       dataLookup.contains(blockHash).pure[F]
     def topoSort(startBlockNumber: Long): F[Vector[Vector[BlockHash]]] =
       topoSortVector.drop(startBlockNumber.toInt).pure[F]
-    def topoSortTail(tailLength: Long): F[Vector[Vector[BlockHash]]] =
-      topoSortVector.takeRight(tailLength.toInt).pure[F]
+    def topoSortTail(tailLength: Int): F[Vector[Vector[BlockHash]]] =
+      topoSortVector.takeRight(tailLength).pure[F]
     def deriveOrdering(startBlockNumber: Long): F[Ordering[BlockMetadata]] =
       topoSort(startBlockNumber).map { topologicalSorting =>
         val order = topologicalSorting.flatten.zipWithIndex.toMap
@@ -73,7 +73,7 @@ final class InMemBlockDagStorage[F[_]: Monad](
                 case (acc, p) =>
                   val currChildren = acc.getOrElse(p, HashSet.empty[BlockHash])
                   acc.updated(p, currChildren + block.blockHash)
-              }
+            }
           )
       _ <- topoSortRef.update(topoSort => TopologicalSortUtil.update(topoSort, 0L, block))
       _ <- latestMessagesRef.update(_.updated(block.sender, block.blockHash))
